move toast-notice to module, remove context
