+++ src/components/frontend-client-base.tsx
@@ -1,6 +1,5 @@
 import { useRouter } from 'next/router'
 import React from 'react'
-import { notify } from 'react-notify-toast'
 import { ThemeProvider } from 'styled-components'
 
 import { ConditonalWrap } from './conditional-wrap'
@@ -14,7 +13,6 @@ import { EntityIdProvider } from '@/contexts/entity-id-context'
 import { InstanceDataProvider } from '@/contexts/instance-context'
 import { LoggedInComponentsProvider } from '@/contexts/logged-in-components'
 import { LoggedInDataProvider } from '@/contexts/logged-in-data-context'
-import { ToastNoticeProvider } from '@/contexts/toast-notice-context'
 import { InstanceData, LoggedInData } from '@/data-types'
 import { FontFix, PrintStylesheet } from '@/helper/css'
 import type { getInstanceDataByLang } from '@/helper/feature-i18n'
@@ -76,8 +74,6 @@ export function FrontendClientBase({
     loggedInComponents,
   ])
 
-  const toastNotice = notify.createShowQueue()
-
   // dev
   //console.dir(initialProps)
 
@@ -90,27 +86,25 @@ export function FrontendClientBase({
         <LoggedInComponentsProvider value={loggedInComponents}>
           <LoggedInDataProvider value={loggedInData}>
             <EntityIdProvider value={entityId || null}>
-              <ToastNoticeProvider value={toastNotice}>
+              <ConditonalWrap
+                condition={!noHeaderFooter}
+                wrapper={(kids) => <HeaderFooter>{kids}</HeaderFooter>}
+              >
                 <ConditonalWrap
-                  condition={!noHeaderFooter}
-                  wrapper={(kids) => <HeaderFooter>{kids}</HeaderFooter>}
+                  condition={!noContainers}
+                  wrapper={(kids) => (
+                    <RelativeContainer>
+                      <MaxWidthDiv showNav={showNav}>
+                        <main>{kids}</main>
+                      </MaxWidthDiv>
+                    </RelativeContainer>
+                  )}
                 >
-                  <ConditonalWrap
-                    condition={!noContainers}
-                    wrapper={(kids) => (
-                      <RelativeContainer>
-                        <MaxWidthDiv showNav={showNav}>
-                          <main>{kids}</main>
-                        </MaxWidthDiv>
-                      </RelativeContainer>
-                    )}
-                  >
-                    {/* should not be necessary…?*/}
-                    {children as JSX.Element}
-                  </ConditonalWrap>
+                  {/* should not be necessary…?*/}
+                  {children as JSX.Element}
                 </ConditonalWrap>
-                <ToastNotice />
-              </ToastNoticeProvider>
+              </ConditonalWrap>
+              <ToastNotice />
             </EntityIdProvider>
           </LoggedInDataProvider>
         </LoggedInComponentsProvider>
